refactor(cart): goods entity의 SelectedOption과의 연결을 @x 표기법으로 명시하며 장바구니 type 분리
