+++ src/entities/cart/cart.types.ts
@@ -1,0 +1,7 @@
+import { SelectedOption } from "@/entities/goods/@x/cart";
+
+export type CartItem = {
+  id: string;
+  name: string;
+  options: SelectedOption[];
+};
