+++ src/entities/goods/@x/cart.ts
@@ -1,0 +1,1 @@
+export type { SelectedOption } from "../goods.types";
